Refactor: changing indexing from 1 to 0. Ref #1

+++ include/todo/task_info.hpp
@@ -3,5 +3,5 @@
 struct TaskInfo {
     std::string title;
     bool done;
-    uint index;
+    size_t index;
 };
